+++ src/stores/__tests__/SettingsStore.spec.ts
@@ -2,14 +2,15 @@ import { describe, it, expect } from 'vitest';
 import { createTestingPinia } from '@pinia/testing';
 import { createPinia, setActivePinia } from 'pinia';
 
-import { useSettingsStore } from '../SettingsStore';
+import { useSettingsStore, type SettingsType } from '../SettingsStore';
 
-const initialState = {
+const initialState: SettingsType = {
   bugzillaApiKey: '',
   jiraApiKey: '',
   showNotifications: false,
   affectsPerPage: 10,
   trackersPerPage: 10,
+  isHidingLabels: false,
 };
 
 // While not used in this file, store below depends on global pinia test instance
@@ -28,27 +29,26 @@ describe('settingsStore', () => {
   it('initializes', () => {
     expect(settingsStore.$state.settings).toEqual(initialState);
   });
+
   it('saves values', () => {
-    settingsStore.save({
+    const settings = {
       bugzillaApiKey: 'beep-beep-who-got-the-keys-to-the-jeep',
       jiraApiKey: 'beep-beep-who-got-the-keys-to-the-jeep',
-      showNotifications: true,
-      affectsPerPage: 10,
-      trackersPerPage: 10,
-    });
-    expect(
-      settingsStore.settings.bugzillaApiKey === 'beep-beep-who-got-the-keys-to-the-jeep',
-    ).toBe(true);
-    expect(
-      settingsStore.settings.jiraApiKey === 'beep-beep-who-got-the-keys-to-the-jeep',
-    ).toBe(true);
-    expect(
-      settingsStore.settings.showNotifications,
-    ).toBe(true);
+      showNotifications: !initialState.showNotifications,
+      affectsPerPage: 1337,
+      trackersPerPage: 1337,
+      isHidingLabels: !initialState.isHidingLabels,
+    };
+
+    settingsStore.save(settings);
+
+    expect(settingsStore.settings).toEqual(settings);
   });
+
   it('reset', () => {
     settingsStore.settings.showNotifications = true;
     settingsStore.$reset();
+
     expect(settingsStore.settings.showNotifications).toBe(false);
   });
 });